Mr/cambios correccions

+++ P2_TPI/src/p2_tpi/Service/GenericService.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface GenericService<T> {
     void insertar(T entidad) throws Exception;
     void actualizar(T entidad) throws Exception;
-    void eliminar(int id) throws Exception;
-    T getById(int id) throws Exception;
+    void eliminar(long id) throws Exception;
+    T getById(long id) throws Exception;
     List<T> getAll() throws Exception;
 }
